# Faster name resolving