--- a/rules/Arguments/Rector/ClassMethod/ReplaceArgumentDefaultValueRector.php
+++ b/rules/Arguments/Rector/ClassMethod/ReplaceArgumentDefaultValueRector.php
@@ -83,8 +83,13 @@ CODE_SAMPLE
             return $this->refactorNew($node);
         }
 
+        $nodeName = $this->getName($node->name);
+        if ($nodeName === null) {
+            return null;
+        }
+
         foreach ($this->replaceArgumentDefaultValues as $replaceArgumentDefaultValue) {
-            if (! $this->isName($node->name, $replaceArgumentDefaultValue->getMethod())) {
+            if (! $this->nodeNameResolver->isStringName($nodeName, $replaceArgumentDefaultValue->getMethod())) {
                 continue;
             }