--- a/rules/Renaming/Rector/MethodCall/RenameMethodRector.php
+++ b/rules/Renaming/Rector/MethodCall/RenameMethodRector.php
@@ -163,22 +163,7 @@ CODE_SAMPLE
             }
 
             foreach ($this->methodCallRenames as $methodCallRename) {
-                if (! $this->nodeNameResolver->isStringName($methodName, $methodCallRename->getOldMethod())) {
-                    continue;
-                }
-
-                if (! $this->nodeTypeResolver->isMethodStaticCallOrClassMethodObjectType(
-                    $classMethod,
-                    $methodCallRename->getObjectType()
-                )) {
-                    continue;
-                }
-
-                if ($this->shouldKeepForParentInterface($methodCallRename, $classReflection)) {
-                    continue;
-                }
-
-                if ($this->hasClassNewClassMethod($classOrInterface, $methodCallRename)) {
+                if ($this->shouldSkipRename($methodName, $classMethod, $methodCallRename, $classReflection, $classOrInterface)) {
                     continue;
                 }
 
@@ -192,6 +177,36 @@ CODE_SAMPLE
         }
 
         return null;
+    }
+
+    private function shouldSkipRename(
+        string $methodName,
+        Node\Stmt\ClassMethod $classMethod,
+        MethodCallRenameInterface $methodCallRename,
+        ClassReflection $classReflection,
+        Class_|Interface_ $classOrInterface
+    ): bool
+    {
+        if (! $this->nodeNameResolver->isStringName($methodName, $methodCallRename->getOldMethod())) {
+            return true;
+        }
+
+        if (! $this->nodeTypeResolver->isMethodStaticCallOrClassMethodObjectType(
+            $classMethod,
+            $methodCallRename->getObjectType()
+        )) {
+            return true;
+        }
+
+        if ($this->shouldKeepForParentInterface($methodCallRename, $classReflection)) {
+            return true;
+        }
+
+        if ($this->hasClassNewClassMethod($classOrInterface, $methodCallRename)) {
+            return true;
+        }
+
+        return false;
     }
 
     private function refactorMethodCallAndStaticCall(