in the worker profiles of mautic I could see these rectors spending a lot of time in repeated name resolving.

I moved the name determination before the loop and used a lower level name-resolver method for comparison.
I did so only in the rectors which were reported in profiles

---

running on mautic..

*before this PR*
```
time php ../rector-src/bin/rector.php --dry-run --no-progress-bar --no-diffs --clear-cache
                                                                                                                        
 [OK] 7 files would have changed (dry-run) by Rector                                                                    
                                                                                                                        
php ../rector-src/bin/rector.php --dry-run --no-progress-bar --no-diffs   478.62s user 8.69s system 873% cpu 55.776 total
```

*with this PR*
```
time php ../rector-src/bin/rector.php --dry-run --no-progress-bar --no-diffs --clear-cache
                                                                                                                        
 [OK] 7 files would have changed (dry-run) by Rector                                                                    
                                                                                                                        
php ../rector-src/bin/rector.php --dry-run --no-progress-bar --no-diffs   396.58s user 9.82s system 817% cpu 49.711 total
```

-> 80 seconds faster