@samsonasik 2023-09-09
iirc, the total time is that should be seen, which 6 seconds faster instead of 80 seconds faster, you may looking using parallel so it collected for all processes time.

That still improvement tho :)