@staabm 2023-09-09
any suggestions regarding the last phpstan error?