@samsonasik 2023-09-09
move complex logic to private method should can reduce complexity