@staabm 2023-09-09
I looked at the wrong method and was wondering what PHPStan is complaining about ;)